requirements.txt
numpy==1.24.4
opencv-python==4.8.1.78
tqdm==4.64.0
torch==1.10.1+cu102
torchvision==0.11.2+cu102
pyyaml==6.0.1
huggingface_hub==0.18.0
safetensors==0.4.0
onnx==1.14.0
onnxruntime==1.14.0
faiss-gpu==1.7.2
timm==0.9.7
scipy==1.5.4
requests==2.27.1
protobuf==3.20.3
psutil==5.9.0
onnxoptimizer==0.2.7
fast-histogram==0.11
ruamel.yaml==0.17.4
grad-cam==1.4.8
pandas==2.0.3
opencv-python-headless